requirements.txt
attrs==23.2.0
boto3==1.34.60
botocore==1.34.60
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
s3transfer==0.10.0
six==1.16.0
urllib3==2.2.1
